+++ core/resources/views/front/themes/theme1.blade.php
@@ -42,7 +42,7 @@
 
                     @if (isset($hero_banner))
                         <div class="col-lg-4 d-none d-lg-block">
-                            <a href="{{ $hero_banner['url1'] }}" class="sright-image">
+                            <a href="{{ $hero_banner['url1'] }}" class="sright-image {{ !empty($hero_banner['hide_text1']) ? 'banner-text-hidden' : '' }}">
                                 <img src="{{ url('/core/public/storage/images/' . $hero_banner['img1']) }}" alt="">
                                 <div class="inner-content">
 
@@ -55,7 +55,7 @@
                                     @endif
                                 </div>
                             </a>
-                            <a href="{{ $hero_banner['url2'] }}" class="sright-image mb-0">
+                            <a href="{{ $hero_banner['url2'] }}" class="sright-image mb-0 {{ !empty($hero_banner['hide_text2']) ? 'banner-text-hidden' : '' }}">
                                 <img src="{{ url('/core/public/storage/images/' . $hero_banner['img2']) }}" alt="">
                                 <div class="inner-content">
                                     @if (isset($hero_banner['subtitle2']))
@@ -193,7 +193,7 @@
             <div class="container ">
                 <div class="row gx-3">
                     <div class="col-md-4">
-                        <a href="{{ $banner_first['firsturl1'] }}" class="genius-banner">
+                        <a href="{{ $banner_first['firsturl1'] }}" class="genius-banner {{ !empty($banner_first['hide_text1']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{ url('/core/public/storage/images/' . $banner_first['img1']) }}" alt="">
                             <div class="inner-content">
                                 @if (isset($banner_first['subtitle1']))
@@ -206,7 +206,7 @@
                         </a>
                     </div>
                     <div class="col-md-4">
-                        <a href="{{ $banner_first['firsturl2'] }}" class="genius-banner">
+                        <a href="{{ $banner_first['firsturl2'] }}" class="genius-banner {{ !empty($banner_first['hide_text2']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{ url('/core/public/storage/images/' . $banner_first['img2']) }}" alt="">
                             <div class="inner-content">
                                 @if (isset($banner_first['subtitle2']))
@@ -219,7 +219,7 @@
                         </a>
                     </div>
                     <div class="col-md-4">
-                        <a href="{{ $banner_first['firsturl3'] }}" class="genius-banner">
+                        <a href="{{ $banner_first['firsturl3'] }}" class="genius-banner {{ !empty($banner_first['hide_text3']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{ url('/core/public/storage/images/' . $banner_first['img3']) }}" alt="">
                             <div class="inner-content">
                                 @if (isset($banner_first['subtitle3']))
@@ -327,7 +327,7 @@
             <div class="container ">
                 <div class="row gx-3">
                     <div class="col-md-4">
-                        <a href="{{ $banner_secend['url1'] }}" class="genius-banner">
+                        <a href="{{ $banner_secend['url1'] }}" class="genius-banner {{ !empty($banner_secend['hide_text1']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{ url('/core/public/storage/images/' . $banner_secend['img1']) }}"
                                 alt="">
                             <div class="inner-content">
@@ -342,7 +342,7 @@
                         </a>
                     </div>
                     <div class="col-md-4">
-                        <a href="{{ $banner_secend['url2'] }}" class="genius-banner">
+                        <a href="{{ $banner_secend['url2'] }}" class="genius-banner {{ !empty($banner_secend['hide_text2']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{ url('/core/public/storage/images/' . $banner_secend['img2']) }}"
                                 alt="">
                             <div class="inner-content">
@@ -357,7 +357,7 @@
                         </a>
                     </div>
                     <div class="col-md-4">
-                        <a href="{{ $banner_secend['url3'] }}" class="genius-banner">
+                        <a href="{{ $banner_secend['url3'] }}" class="genius-banner {{ !empty($banner_secend['hide_text3']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{ url('/core/public/storage/images/' . $banner_secend['img3']) }}"
                                 alt="">
                             <div class="inner-content">
@@ -610,7 +610,7 @@
             <div class="container ">
                 <div class="row gx-3">
                     <div class="col-md-6">
-                        <a href="{{ $banner_third['url1'] }}" class="genius-banner">
+                        <a href="{{ $banner_third['url1'] }}" class="genius-banner {{ !empty($banner_third['hide_text1']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{ url('/core/public/storage/images/' . $banner_third['img1']) }}"
                                 alt="">
                             <div class="inner-content">
@@ -624,7 +624,7 @@
                         </a>
                     </div>
                     <div class="col-md-6">
-                        <a href="{{ $banner_third['url2'] }}" class="genius-banner">
+                        <a href="{{ $banner_third['url2'] }}" class="genius-banner {{ !empty($banner_third['hide_text2']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{ url('/core/public/storage/images/' . $banner_third['img2']) }}"
                                 alt="">
                             <div class="inner-content">