+++ core/resources/views/front/themes/theme3.blade.php
@@ -84,7 +84,7 @@
             <div class="container ">
                 <div class="row gx-3">
                     <div class="col-md-4">
-                        <a href="{{$banner_first['firsturl1']}}" class="genius-banner">
+                        <a href="{{$banner_first['firsturl1']}}" class="genius-banner {{ !empty($banner_first['hide_text1']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{ url('/core/public/storage/images/'.$banner_first['img1']) }}" alt="">
                             <div class="inner-content">
                                 @if (isset($banner_first['subtitle1']))
@@ -97,7 +97,7 @@
                         </a>
                     </div>
                     <div class="col-md-4">
-                        <a href="{{$banner_first['firsturl2']}}" class="genius-banner">
+                        <a href="{{$banner_first['firsturl2']}}" class="genius-banner {{ !empty($banner_first['hide_text2']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{ url('/core/public/storage/images/'.$banner_first['img2']) }}" alt="">
                             <div class="inner-content">
                                 @if (isset($banner_first['subtitle2']))
@@ -110,7 +110,7 @@
                         </a>
                     </div>
                     <div class="col-md-4">
-                        <a href="{{$banner_first['firsturl3']}}" class="genius-banner">
+                        <a href="{{$banner_first['firsturl3']}}" class="genius-banner {{ !empty($banner_first['hide_text3']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{ url('/core/public/storage/images/'.$banner_first['img3']) }}" alt="">
                             <div class="inner-content">
                                 @if (isset($banner_first['subtitle3']))
@@ -275,7 +275,7 @@
         <div class="container ">
             <div class="row gx-3">
                 <div class="col-md-4">
-                    <a href="{{$banner_secend['url1']}}" class="genius-banner">
+                    <a href="{{$banner_secend['url1']}}" class="genius-banner {{ !empty($banner_secend['hide_text1']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_secend['img1']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_secend['subtitle1']))
@@ -289,7 +289,7 @@
                     </a>
                 </div>
                 <div class="col-md-4">
-                    <a href="{{$banner_secend['url2']}}" class="genius-banner">
+                    <a href="{{$banner_secend['url2']}}" class="genius-banner {{ !empty($banner_secend['hide_text2']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_secend['img2']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_secend['subtitle2']))
@@ -303,7 +303,7 @@
                     </a>
                 </div>
                 <div class="col-md-4">
-                    <a href="{{$banner_secend['url3']}}" class="genius-banner">
+                    <a href="{{$banner_secend['url3']}}" class="genius-banner {{ !empty($banner_secend['hide_text3']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_secend['img3']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_secend['subtitle3']))
@@ -449,7 +449,7 @@
         <div class="container ">
             <div class="row gx-3">
                 <div class="col-md-6">
-                    <a href="{{$banner_third['url1']}}" class="genius-banner">
+                    <a href="{{$banner_third['url1']}}" class="genius-banner {{ !empty($banner_third['hide_text1']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_third['img1']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_third['subtitle1']))
@@ -462,7 +462,7 @@
                     </a>
                 </div>
                 <div class="col-md-6">
-                    <a href="{{$banner_third['url2']}}" class="genius-banner">
+                    <a href="{{$banner_third['url2']}}" class="genius-banner {{ !empty($banner_third['hide_text2']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_third['img2']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_third['subtitle2']))