+++ core/resources/views/front/themes/theme4.blade.php
@@ -27,27 +27,27 @@
             <div class="container">
                 <div class="row">
                     <div class="col-lg-3 col-md-3 col-sm-12">
-                        <a href="{{isset($home_page4_banner['url1']) ? $home_page4_banner['url1'] : ''}}" class="h3-category">
+                        <a href="{{isset($home_page4_banner['url1']) ? $home_page4_banner['url1'] : ''}}" class="h3-category {{ !empty($home_page4_banner['hide_text1']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{url('/core/public/storage/images/'.$home_page4_banner['img1'])}}" alt="">
                             <h4>{{isset($home_page4_banner['label1']) ? $home_page4_banner['label1'] : ''}}</h4>
                         </a>
-                        <a href="{{isset($home_page4_banner['url2']) ? $home_page4_banner['url2'] : ''}}" class="h3-category">
+                        <a href="{{isset($home_page4_banner['url2']) ? $home_page4_banner['url2'] : ''}}" class="h3-category {{ !empty($home_page4_banner['hide_text2']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{url('/core/public/storage/images/'.$home_page4_banner['img2'])}}" alt="">
                             <h4>{{isset($home_page4_banner['label2']) ? $home_page4_banner['label2'] : ''}}</h4>
                         </a>
                     </div>
                     <div class="col-lg-6 col-md-6 col-sm-12 middleImage">
-                        <a href="{{isset($home_page4_banner['url3']) ? $home_page4_banner['url3'] : ''}}" class="h3-category">
+                        <a href="{{isset($home_page4_banner['url3']) ? $home_page4_banner['url3'] : ''}}" class="h3-category {{ !empty($home_page4_banner['hide_text3']) ? 'banner-text-hidden' : '' }}">
                             <img class="lazy" data-src="{{url('/core/public/storage/images/'.$home_page4_banner['img3'])}}" alt="">
                             <h4>{{isset($home_page4_banner['label3']) ? $home_page4_banner['label3'] : ''}}</h4>
                         </a>
                     </div>
                     <div class="col-lg-3 col-md-3 col-sm-12">
-                        <a href="{{isset($home_page4_banner['url4']) ? $home_page4_banner['url4'] : ''}}" class="h3-category">
+                        <a href="{{isset($home_page4_banner['url4']) ? $home_page4_banner['url4'] : ''}}" class="h3-category {{ !empty($home_page4_banner['hide_text4']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{url('/core/public/storage/images/'.$home_page4_banner['img4'])}}" alt="">
                             <h4>{{isset($home_page4_banner['label4']) ? $home_page4_banner['label4'] : ''}}</h4>
                         </a>
-                        <a href="{{isset($home_page4_banner['url5']) ? $home_page4_banner['url5'] : ''}}" class="h3-category">
+                        <a href="{{isset($home_page4_banner['url5']) ? $home_page4_banner['url5'] : ''}}" class="h3-category {{ !empty($home_page4_banner['hide_text5']) ? 'banner-text-hidden' : '' }}">
                             <img src="{{url('/core/public/storage/images/'.$home_page4_banner['img5'])}}" alt="">
                             <h4>{{isset($home_page4_banner['label5']) ? $home_page4_banner['label5'] : ''}}</h4>
                         </a>
@@ -136,7 +136,7 @@
         <div class="container ">
             <div class="row gx-3">
                 <div class="col-md-4">
-                    <a href="{{$banner_first['firsturl1']}}" class="genius-banner">
+                    <a href="{{$banner_first['firsturl1']}}" class="genius-banner {{ !empty($banner_first['hide_text1']) ? 'banner-text-hidden' : '' }}">
                         <img src="{{ url('/core/public/storage/images/'.$banner_first['img1']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_first['subtitle1']))
@@ -149,7 +149,7 @@
                     </a>
                 </div>
                 <div class="col-md-4">
-                    <a href="{{$banner_first['firsturl2']}}" class="genius-banner">
+                    <a href="{{$banner_first['firsturl2']}}" class="genius-banner {{ !empty($banner_first['hide_text2']) ? 'banner-text-hidden' : '' }}">
                         <img src="{{ url('/core/public/storage/images/'.$banner_first['img2']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_first['subtitle2']))
@@ -162,7 +162,7 @@
                     </a>
                 </div>
                 <div class="col-md-4">
-                    <a href="{{$banner_first['firsturl3']}}" class="genius-banner">
+                    <a href="{{$banner_first['firsturl3']}}" class="genius-banner {{ !empty($banner_first['hide_text3']) ? 'banner-text-hidden' : '' }}">
                         <img src="{{ url('/core/public/storage/images/'.$banner_first['img3']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_first['subtitle3']))
@@ -253,7 +253,7 @@
         <div class="container ">
             <div class="row gx-3">
                 <div class="col-md-4">
-                    <a href="{{$banner_secend['url1']}}" class="genius-banner">
+                    <a href="{{$banner_secend['url1']}}" class="genius-banner {{ !empty($banner_secend['hide_text1']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_secend['img1']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_secend['subtitle1']))
@@ -267,7 +267,7 @@
                     </a>
                 </div>
                 <div class="col-md-4">
-                    <a href="{{$banner_secend['url2']}}" class="genius-banner">
+                    <a href="{{$banner_secend['url2']}}" class="genius-banner {{ !empty($banner_secend['hide_text2']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_secend['img2']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_secend['subtitle2']))
@@ -281,7 +281,7 @@
                     </a>
                 </div>
                 <div class="col-md-4">
-                    <a href="{{$banner_secend['url3']}}" class="genius-banner">
+                    <a href="{{$banner_secend['url3']}}" class="genius-banner {{ !empty($banner_secend['hide_text3']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_secend['img3']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_secend['subtitle3']))
@@ -387,7 +387,7 @@
         <div class="container ">
             <div class="row gx-3">
                 <div class="col-md-6">
-                    <a href="{{$banner_third['url1']}}" class="genius-banner">
+                    <a href="{{$banner_third['url1']}}" class="genius-banner {{ !empty($banner_third['hide_text1']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_third['img1']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_third['subtitle1']))
@@ -400,7 +400,7 @@
                     </a>
                 </div>
                 <div class="col-md-6">
-                    <a href="{{$banner_third['url2']}}" class="genius-banner">
+                    <a href="{{$banner_third['url2']}}" class="genius-banner {{ !empty($banner_third['hide_text2']) ? 'banner-text-hidden' : '' }}">
                         <img class="lazy" data-src="{{ url('/core/public/storage/images/'.$banner_third['img2']) }}" alt="">
                         <div class="inner-content">
                             @if (isset($banner_third['subtitle2']))